Add vote preference stakepool integration.
This updates the stakepool client to support the v2 API which can be
used to set vote preferences, and updates to a newer version of the
wallet gRPC API to support saving and retreiving vote preferences from
the wallet DB.

All agendas in the latest support stake version are now shown in the
purchase tickets view with dropdown boxes for users to select their
vote preference on an agenda.  Whenever a choice is changed, the local
vote preferences are saved by the wallet and each configured
stakepool is updated through the API.

Note that vote preferences are never read through the stakepool API.
If a user changes their preferences through another channel (e.g. the
stakepool web UI, or other wallet software) the update will not be
visible in Paymetheus and Paymetheus may overwrite preferences when
additional modifications are made.  This is unfortunate but I don't
see a better way of handling this considering that multiple stakepools
may be configured, each with different vote preferences.

Closes #240.