Add checkbox to rescan for imported scripts.
The rescan starts from the genesis block and occurs in the background
and continues after the dialog closes.  If there is an error, an error
message is opened, but otherwise there is no feedback about when the
rescan is complete.  Such a feature would have to be integrated into
the shell viewmodel.  For now, just make the feature available.