Add button to create ticket revocations.
This button has been added to the "Purchase tickets" view, which is
know just called "Tickets" for simplicity.

The navigation in the shell view has been reorganized, and the Tickets
view is now beside the "Stake mining" view (which is planned to be
folded into the Tickets view at a later time).

There is not yet any indication of whether revocations must be created
by the user, but this is the first step to making the wallet more
usable by being able to unlock missed/expired tickets that would
otherwise remain indefinitely unspendable.