Improve UX of the purchase ticket view.
This change removes the split tx (this can be always be left as the
default, and should use a configured relay fee instead of being a
separate option) and renames the "Ticket fee" field to "Ticket
transaction fee" to make it clearer that it is a transaction fee and
is unrelated to the ticket price.

While here, drastically improve the XAML for the view by removing an
unnecessary grid column, unnecessary grid and row spans, unnecessary
margins and offsets that should be handled by the grid itself, and
inconsistent styling of elements.