Verify stakepool multisig script and p2sh voting addresses.
This check ensures that the stakepool 1-of-2 multisig script can be
redeemed by the user's wallet (preventing stakepools from extorting
users for votes and revocations) and verifies that the address given
ticket voting rights to is the P2SH address of the same multisig
script.

When a possible bug or malicious stakepool behavior is discovered, a
warning message is presented to the user and the program will refuse
to purchase tickets.